cleanup loadmodels
