+++ src/context/SceneContext.jsx
@@ -1,45 +1,11 @@
 import React, { createContext, useEffect, useState } from "react"
 import * as THREE from "three"
-import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader"
-import { VRMLoaderPlugin } from "@pixiv/three-vrm"
 import { cullHiddenMeshes } from "../library/utils"
-import {
-  renameVRMBones,
-  createFaceNormals,
-  createBoneDirection,
-} from "../library/utils"
 
 export const SceneContext = createContext()
 
 export const SceneProvider = (props) => {
-  const loadingManager = new THREE.LoadingManager()
 
-  const gltfLoader = new GLTFLoader(loadingManager)
-  gltfLoader.register((parser) => {
-    return new VRMLoaderPlugin(parser)
-  })
-
-  async function loadModel(file, onProgress) {
-    return gltfLoader.loadAsync(file, onProgress).then((model) => {
-      return addModel(model);
-    })
-  }
-
-  // separated to call it after load manager finishes
-  function addModel(model){
-    const vrm = model.userData.vrm
-    renameVRMBones(vrm)
-
-    vrm.scene?.traverse((child) => {
-      child.frustumCulled = false
-
-      if (child.isMesh) {
-        createFaceNormals(child.geometry)
-        if (child.isSkinnedMesh) createBoneDirection(child)
-      }
-    })
-    return vrm
-  }
   function getAsArray(target){
     if (target == null) return []
     return Array.isArray(target) ? target : [target]
@@ -91,12 +57,10 @@ export const SceneProvider = (props) => {
         setCurrentTraitName,
         currentOptions,
         setCurrentOptions,
-        loadModel,
         setSelectedOptions,
         selectedOptions,
         setSelectedRandomTraits,
         selectedRandomTraits,
-        addModel,
         model,
         setModel,
         animationManager,
